+++ src/onegov.org/HISTORY.rst
@@ -1,6 +1,9 @@
 Changelog
 ---------
 
+1.3.0 (2019-05-16)
+~~~~~~~~~~~~~~~~~~~
+
 - Warns users about to close a ticket without making a decision.
   [href]